First test passes; on to the others
